Fixed bug in create_reduced_variant_files.R
Discovered that the bug that was in this file was due to the command line argument of a number be automatically converted to a string. Of course this does not happen in RStudio, and so that is why I kept getting different numbers when trying to subset with  "7" instead of 7.